project/requirements.txt
flask
mysql-connector-python
pandas
openpyxl
jinja2
python-dotenv
